Split Demographic Test into its own file
Also have the `DemographicsService` use the `DemogrphicsService` config for the S3 bucket location
